+++ pynotify/config.py
@@ -3,13 +3,14 @@ from django.conf import settings as django_settings
 
 class Settings:
     """
-    Holds default configuration values, the values can be overriden in settings with ``PYNOTIFY_`` prefix.
+    Holds default configuration values, the values can be overridden in settings with ``PYNOTIFY_`` prefix.
     """
     PREFIX = 'PYNOTIFY'
     DEFAULTS = {
         'AUTOLOAD_MODULES': None,
         'CELERY_TASK': 'pynotify.tasks.notification_task',
         'ENABLED': True,
+        'NOTIFICATION_MODEL': 'pynotify.Notification',
         'RECEIVER': 'pynotify.receivers.SynchronousReceiver',
         'RELATED_OBJECTS_ALLOWED_ATTRIBUTES': {'get_absolute_url', },
         'STRIP_HTML': False,